Develop/Flask_api/requirements.txt
jieba==0.42.1
PyMySQL==0.9.3
Flask==1.1.2
selenium==3.141.0
google_trans_new==1.1.9
python_wordpress_xmlrpc==2.3
requests==2.22.0
redis==3.4.1
beautifulsoup4==4.9.3
lxml
